requirements.txt
python-escpos
requests
flask
python-dotenv